Deprecate "core/parsers" folder
- deprecate notice for "core/parsers" folder (will be removed after v3.4)
- move entire parsing logic within "core/utils/parsers"
