+++ src/app/core/parsers/table/parser.js
@@ -1,29 +1,6 @@
-const Feature = require('core/layers/features/feature');
-const TableParser = function() {
-  this.get = function(options={}) {
-    const type = options.type;
-    let parser;
-    switch (type) {
-      case 'json':
-        parser = this._parserJSON.bind(this);
-        break;
-      default:
-        parser = this._parserJSON.bind(this);
-    }
-    return parser;
-  };
+const { table } = require('core/utils/parsers');
 
-  this._parserJSON = function(data={}) {
-    const {features=[]} = data;
-    return features.map(_feature => {
-      const {id, properties} = _feature;
-      const feature = new Feature();
-      feature.setProperties(properties);
-      feature.setId(id);
-      return feature;
-    });
-  }
-};
-
-module.exports = new TableParser();
-
+/**
+ * DEPRECATED: this folder will be removed after v3.4 (use "core/utils/parsers" instead)
+ */
+module.exports = table;
