+++ src/app/gui/queryresults/vue/components/actiontools/infoformats.vue
@@ -6,7 +6,7 @@
 
 <script>
     const CatalogLayersStoresRegistry = require('core/catalog/cataloglayersstoresregistry');
-    const responseParser = require('core/parsers/response/parser');
+    const {response: responseParser} = require('core/utils/parsers');
     const {getAlphanumericPropertiesFromFeature, query} = require('core/utils/geo');
     const GUI = require('gui/gui');
     export default {
